refactoring. added shared methods to DiskBasedCacheUtility.
