+++ src/main/java/com/android/volley/toolbox/DiskBasedAsyncCache.java
@@ -3,7 +3,6 @@ package com.android.volley.toolbox;
 import android.annotation.SuppressLint;
 import androidx.annotation.VisibleForTesting;
 import com.android.volley.AsyncCache;
-import com.android.volley.Header;
 import java.io.File;
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -12,9 +11,7 @@ import java.nio.channels.CompletionHandler;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
-import java.util.Collections;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -25,14 +22,13 @@ import java.util.Map;
 public class DiskBasedAsyncCache extends AsyncCache {
 
     /** Map of the Key, CacheHeader pairs */
-    private final Map<String, DiskBasedAsyncCache.CacheHeader> mEntries =
-            new LinkedHashMap<>(16, .75f, true);
+    private final Map<String, CacheHeader> mEntries = new LinkedHashMap<>(16, .75f, true);
 
     /** Total amount of space currently used by the cache in bytes. */
     private long mTotalSize = 0;
 
     /** The supplier for the root directory to use for the cache. */
-    private final DiskBasedAsyncCache.FileSupplier mRootDirectorySupplier;
+    private final DiskBasedCacheUtility.FileSupplier mRootDirectorySupplier;
 
     /** The maximum size of the cache in bytes. */
     private final int mMaxCacheSizeInBytes;
@@ -56,7 +52,7 @@ public class DiskBasedAsyncCache extends AsyncCache {
      */
     public DiskBasedAsyncCache(final File rootDirectory, int maxCacheSizeInBytes) {
         mRootDirectorySupplier =
-                new DiskBasedAsyncCache.FileSupplier() {
+                new DiskBasedCacheUtility.FileSupplier() {
                     @Override
                     public File get() {
                         return rootDirectory;
@@ -75,7 +71,7 @@ public class DiskBasedAsyncCache extends AsyncCache {
     @Override
     public void get(String key, OnGetCompleteCallback callback) {
         final OnGetCompleteCallback cb = callback;
-        final DiskBasedAsyncCache.CacheHeader entry = mEntries.get(key);
+        final CacheHeader entry = mEntries.get(key);
         // if the entry does not exist, return.
         if (entry == null) {
             cb.onGetComplete(null);
@@ -142,118 +138,8 @@ public class DiskBasedAsyncCache extends AsyncCache {
         // TODO (sphill99): Implement
     }
 
-    /**
-     * Creates a pseudo-unique filename for the specified cache key.
-     *
-     * @param key The key to generate a file name for.
-     * @return A pseudo-unique filename.
-     */
-    private String getFilenameForKey(String key) {
-        int firstHalfLength = key.length() / 2;
-        String localFilename = String.valueOf(key.substring(0, firstHalfLength).hashCode());
-        localFilename += String.valueOf(key.substring(firstHalfLength).hashCode());
-        return localFilename;
-    }
-
     /** Returns a file object for the given cache key. */
     public File getFileForKey(String key) {
-        return new File(mRootDirectorySupplier.get(), getFilenameForKey(key));
-    }
-
-    /** Represents a supplier for {@link File}s. */
-    public interface FileSupplier {
-        File get();
-    }
-
-    /** Handles holding onto the cache headers for an entry. */
-    @VisibleForTesting
-    static class CacheHeader {
-        /**
-         * The size of the data identified by this CacheHeader on disk (both header and data).
-         *
-         * <p>Must be set by the caller after it has been calculated.
-         *
-         * <p>This is not serialized to disk.
-         */
-        long size;
-
-        /** The key that identifies the cache entry. */
-        final String key;
-
-        /** ETag for cache coherence. */
-        final String etag;
-
-        /** Date of this response as reported by the server. */
-        final long serverDate;
-
-        /** The last modified date for the requested object. */
-        final long lastModified;
-
-        /** TTL for this record. */
-        final long ttl;
-
-        /** Soft TTL for this record. */
-        final long softTtl;
-
-        /** Headers from the response resulting in this cache entry. */
-        final List<Header> allResponseHeaders;
-
-        private CacheHeader(
-                String key,
-                String etag,
-                long serverDate,
-                long lastModified,
-                long ttl,
-                long softTtl,
-                List<Header> allResponseHeaders) {
-            this.key = key;
-            this.etag = "".equals(etag) ? null : etag;
-            this.serverDate = serverDate;
-            this.lastModified = lastModified;
-            this.ttl = ttl;
-            this.softTtl = softTtl;
-            this.allResponseHeaders = allResponseHeaders;
-        }
-
-        /**
-         * Instantiates a new CacheHeader object.
-         *
-         * @param key The key that identifies the cache entry
-         * @param entry The cache entry.
-         */
-        CacheHeader(String key, Entry entry) {
-            this(
-                    key,
-                    entry.etag,
-                    entry.serverDate,
-                    entry.lastModified,
-                    entry.ttl,
-                    entry.softTtl,
-                    getAllResponseHeaders(entry));
-        }
-
-        private static List<Header> getAllResponseHeaders(Entry entry) {
-            // If the entry contains all the response headers, use that field directly.
-            if (entry.allResponseHeaders != null) {
-                return entry.allResponseHeaders;
-            }
-
-            // Legacy fallback - copy headers from the map.
-            return HttpHeaderParser.toAllHeaderList(entry.responseHeaders);
-        }
-
-        /** Creates a cache entry for the specified data. */
-        Entry toCacheEntry(byte[] data) {
-            Entry e = new Entry();
-            e.data = data;
-            e.etag = etag;
-            e.serverDate = serverDate;
-            e.lastModified = lastModified;
-            e.ttl = ttl;
-            e.softTtl = softTtl;
-            e.responseHeaders = HttpHeaderParser.toHeaderMap(allResponseHeaders);
-            e.allResponseHeaders = Collections.unmodifiableList(allResponseHeaders);
-            return e;
-        }
+        return new File(mRootDirectorySupplier.get(), DiskBasedCacheUtility.getFilenameForKey(key));
     }
 }
